contact email working, background set to take full width, nav links working...hash hash baby
