+++ profile/public/index.html
@@ -30,7 +30,6 @@
   </head>
   <body class="d-flex flex-column h-100
   ">
-    <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
     <!--
       This HTML file is a template.
