+++ profile/src/Components/Pages/Contact/Contact.css
@@ -1,0 +1,3 @@
+.container{
+    height: 100vh
+}
