+++ profile/src/Components/Pages/Cover/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Cover";
