+++ profile/src/Components/Pages/Profile/Profile.css
@@ -1,0 +1,3 @@
+.container{
+    height: 100vh
+}
