+++ profile/src/Components/Pages/Projects/Projects.css
@@ -1,0 +1,3 @@
+.container{
+    height: 100vh
+}
