+++ profile/src/Components/Pages/Projects/index.js
@@ -1,0 +1,1 @@
+export {default} from "./Projects";
