+++ profile/src/Components/Pages/Skills/Skills.css
@@ -1,0 +1,3 @@
+.container{
+    height: 100vh
+}
