+++ profile/src/Components/Pages/Skills/index.js
@@ -1,0 +1,1 @@
+export {default} from './Skills';
