--- profile/src/Pages/Cover/Cover.css
@@ -1,29 +1,0 @@
-.carousel-item {
-    height: 100vh;
-    min-height: 350px;
-    background: no-repeat center center scroll;
-    -webkit-background-size: cover;
-    -moz-background-size: cover;
-    -o-background-size: cover;
-    background-size: cover;
-  }
-
-  .overlay { 
-    color:#fff;
-    position:absolute;
-    z-index:12;
-    left:0;
-    right:0;
-    width:100%;
-    text-align:center;
-  }
-
-
-  .downButton{
-    position: absolute;
-    bottom: 0%;
-    right: 0%
-  }
-
-  
-
