--- profile/src/Pages/Experience/Experience.css
@@ -1,3 +1,0 @@
-.container{
-    height: 100vh
-}
