--- profile/src/Pages/Experience/index.js
@@ -1,1 +1,0 @@
-export {default} from "./Experience";
