--- profile/src/Pages/Skills/index.js
@@ -1,1 +1,0 @@
-export {default} from './Skills';
